task-03-server-side-tools/app/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
confluent-kafka==2.12.2
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.124.4
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
idna==3.11
importlib-metadata==8.7.0
loki-logger-handler==1.1.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
packaging==25.0
prometheus-client==0.23.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-core==2.41.5
python-dateutil==2.9.0.post0
pyyaml==6.0.3
redis==7.1.0
requests==2.32.5
seqlog==0.4.3
six==1.17.0
starlette==0.50.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.2
uvicorn==0.38.0
wrapt==1.17.3
zipp==3.23.0
